requirements.txt
flask==3.1.3
requests==2.32.5
python-dotenv==1.2.2